+++ stories/documentation/structure/box/box-basic.stories.ts
@@ -1,11 +1,11 @@
 import { Meta, StoryFn } from '@storybook/angular';
 
-interface BoxToggleStory {
+interface BoxBasicStory {
 	grey: boolean;
 }
 
 export default {
-	title: 'Documentation/Feedback/Box/Toggle',
+	title: 'Documentation/Structure/Box/Basic',
 	argTypes: {
 		grey: {
 			control: {
@@ -15,21 +15,17 @@ export default {
 	},
 } as Meta;
 
-function getTemplate(args: BoxToggleStory): string {
+function getTemplate(args: BoxBasicStory): string {
 	const grey = args.grey ? `mod-grey` : '';
 
 	return `
-	<div class="switch">
-	    <input class="switch-input" type="checkbox" id="boxSwitch" checked disabled>
-	    <label class="switch-label" for="boxSwitch">Switch</label>
-	</div>
-	<div class="box mod-toggle ${grey}">
+	<div class="box ${grey}">
 	    Jujubes toppin gvueoat cake cake lemon drops chupa chups sweet roll. Macaroon icing tootsie roll bonbon dragée carrot cake sweet roll. Pie gingerbread jelly beans cotton candy tart lollipop bonbon candy. Bonbon chocolate gingerbread pastry.
 	</div>
 	`;
 }
 
-const Template: StoryFn<BoxToggleStory> = (args: BoxToggleStory) => ({
+const Template: StoryFn<BoxBasicStory> = (args: BoxBasicStory) => ({
 	props: args,
 	template: getTemplate(args),
 	styles: [
@@ -41,5 +37,5 @@ const Template: StoryFn<BoxToggleStory> = (args: BoxToggleStory) => ({
 	],
 });
 
-export const Toggle = Template.bind({});
-Toggle.args = { grey: false };
+export const Basic = Template.bind({});
+Basic.args = { grey: false };
