+++ stories/documentation/structure/box/box-toggle.stories.ts
@@ -1,11 +1,11 @@
 import { Meta, StoryFn } from '@storybook/angular';
 
-interface BoxBasicStory {
+interface BoxToggleStory {
 	grey: boolean;
 }
 
 export default {
-	title: 'Documentation/Feedback/Box/Basic',
+	title: 'Documentation/Structure/Box/Toggle',
 	argTypes: {
 		grey: {
 			control: {
@@ -15,17 +15,21 @@ export default {
 	},
 } as Meta;
 
-function getTemplate(args: BoxBasicStory): string {
+function getTemplate(args: BoxToggleStory): string {
 	const grey = args.grey ? `mod-grey` : '';
 
 	return `
-	<div class="box ${grey}">
+	<div class="switch">
+	    <input class="switch-input" type="checkbox" id="boxSwitch" checked disabled>
+	    <label class="switch-label" for="boxSwitch">Switch</label>
+	</div>
+	<div class="box mod-toggle ${grey}">
 	    Jujubes toppin gvueoat cake cake lemon drops chupa chups sweet roll. Macaroon icing tootsie roll bonbon dragée carrot cake sweet roll. Pie gingerbread jelly beans cotton candy tart lollipop bonbon candy. Bonbon chocolate gingerbread pastry.
 	</div>
 	`;
 }
 
-const Template: StoryFn<BoxBasicStory> = (args: BoxBasicStory) => ({
+const Template: StoryFn<BoxToggleStory> = (args: BoxToggleStory) => ({
 	props: args,
 	template: getTemplate(args),
 	styles: [
@@ -37,5 +41,5 @@ const Template: StoryFn<BoxBasicStory> = (args: BoxBasicStory) => ({
 	],
 });
 
-export const Basic = Template.bind({});
-Basic.args = { grey: false };
+export const Toggle = Template.bind({});
+Toggle.args = { grey: false };
